+++ src/Thrift.Net.Compilation/Symbols/CollectionType.cs
@@ -1,9 +1,6 @@
 namespace Thrift.Net.Compilation.Symbols
 {
-    using System;
-    using System.Collections.Generic;
     using Antlr4.Runtime.Tree;
-    using Thrift.Net.Compilation.Binding;
 
     /// <summary>
     /// A base class for symbols that represent collections, like list or set.
@@ -12,48 +9,16 @@ namespace Thrift.Net.Compilation.Symbols
     public abstract class CollectionType<TNode> : Symbol<TNode, ISymbol>, ICollectionType
         where TNode : IParseTree
     {
-        private readonly IBinderProvider binderProvider;
-        private readonly string collectionTypeName;
-        private readonly string csharpCollectionTypeName;
-
         /// <summary>
         /// Initializes a new instance of the <see cref="CollectionType{TNode}" /> class.
         /// </summary>
         /// <param name="node">The node this symbol is bound to.</param>
         /// <param name="parent">The parent symbol.</param>
-        /// <param name="binderProvider">Used to get binders for nodes.</param>
-        /// <param name="collectionTypeName">The name of the Thrift collection type (e.g. `list`).</param>
-        /// <param name="csharpCollectionTypeName">The name of the C# type that will be generated from this type.</param>
         protected CollectionType(
             TNode node,
-            ISymbol parent,
-            IBinderProvider binderProvider,
-            string collectionTypeName,
-            string csharpCollectionTypeName)
+            ISymbol parent)
             : base(node, parent)
         {
-            this.binderProvider = binderProvider;
-            this.collectionTypeName = collectionTypeName;
-            this.csharpCollectionTypeName = csharpCollectionTypeName;
-        }
-
-        /// <summary>
-        /// Gets the type of the collection's element.
-        /// </summary>
-        public IFieldType ElementType
-        {
-            get
-            {
-                var elementNode = this.GetElementNode();
-                if (elementNode != null)
-                {
-                    return this.binderProvider
-                        .GetBinder(elementNode)
-                        .Bind<IFieldType>(elementNode, this);
-                }
-
-                return null;
-            }
         }
 
         /// <inheritdoc/>
@@ -79,7 +44,7 @@ namespace Thrift.Net.Compilation.Symbols
         }
 
         /// <inheritdoc/>
-        public string Name => $"{this.collectionTypeName}<{this.ElementType?.Name}>";
+        public string Name => this.Node.GetText();
 
         /// <inheritdoc/>
         public bool IsResolved => true;
@@ -109,26 +74,14 @@ namespace Thrift.Net.Compilation.Symbols
         public bool IsSet => this is ISetType;
 
         /// <inheritdoc/>
-        public bool IsMap => false;
-
-        /// <inheritdoc/>
-        protected override IReadOnlyCollection<ISymbol> Children =>
-            this.ElementType != null ? new List<ISymbol> { this.ElementType } : new List<ISymbol>();
+        public bool IsMap => this is IMapType;
 
         /// <summary>
-        /// Gets the Antlr tree node representing the collection's element type.
+        /// Gets the C# type name for the collection.
         /// </summary>
-        /// <returns>The element node.</returns>
-        protected abstract IParseTree GetElementNode();
-
-        private string GetTypeName()
-        {
-            if (this.ElementType != null)
-            {
-                return $"{this.csharpCollectionTypeName}<{this.ElementType.CSharpRequiredTypeName}>";
-            }
-
-            throw new InvalidOperationException("Cannot get the type name because no element type was provided.");
-        }
+        /// <returns>
+        /// The type name.
+        /// </returns>
+        protected abstract string GetTypeName();
     }
 }
