+++ src/Thrift.Net.Compilation/Symbols/MapType.cs
@@ -7,7 +7,7 @@ namespace Thrift.Net.Compilation.Symbols
     /// <summary>
     /// Represents a Thrift map.
     /// </summary>
-    public class MapType : Symbol<MapTypeContext, ISymbol>, IMapType
+    public class MapType : CollectionType<MapTypeContext>, IMapType
     {
         private readonly IBinderProvider binderProvider;
 
@@ -56,62 +56,6 @@ namespace Thrift.Net.Compilation.Symbols
         }
 
         /// <inheritdoc/>
-        public string Name => this.Node.GetText();
-
-        /// <inheritdoc/>
-        public bool IsResolved => true;
-
-        /// <inheritdoc/>
-        public string CSharpOptionalTypeName => this.GetCSharpTypeName();
-
-        /// <inheritdoc/>
-        public string CSharpRequiredTypeName => this.GetCSharpTypeName();
-
-        /// <inheritdoc/>
-        public bool IsBaseType => false;
-
-        /// <inheritdoc/>
-        public bool IsStruct => false;
-
-        /// <inheritdoc/>
-        public bool IsEnum => false;
-
-        /// <inheritdoc/>
-        public bool IsList => false;
-
-        /// <inheritdoc/>
-        public bool IsCollection => true;
-
-        /// <inheritdoc/>
-        public bool IsSet => false;
-
-        /// <inheritdoc/>
-        public bool IsMap => true;
-
-        /// <inheritdoc/>
-        public int? NestingDepth
-        {
-            // TODO: Add tests / rejig the inheritance chain so this can use CollectionType
-            get
-            {
-                if (this.Parent is ICollectionType)
-                {
-                    var depth = 1;
-                    var parent = this.Parent;
-                    while (parent.Parent is ICollectionType)
-                    {
-                        parent = parent.Parent;
-                        depth++;
-                    }
-
-                    return depth;
-                }
-
-                return null;
-            }
-        }
-
-        /// <inheritdoc/>
         protected override IReadOnlyCollection<ISymbol> Children
         {
             get
@@ -138,7 +82,8 @@ namespace Thrift.Net.Compilation.Symbols
             base.Accept(visitor);
         }
 
-        private string GetCSharpTypeName()
+        /// <inheritdoc/>
+        protected override string GetTypeName()
         {
             if (this.KeyType != null && this.ValueType != null)
             {
